+++ app/src/main/java/com/naver/maps/map/demo/kotlin/clustering/ComplexClusteringActivity.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018-2024 NAVER Corp.
+ * Copyright 2018-2025 NAVER Corp.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -73,7 +73,6 @@ class ComplexClusteringActivity : AppCompatActivity(), OnMapReadyCallback {
     private var clusterer: Clusterer<ItemKey>? = null
 
     override fun onMapReady(naverMap: NaverMap) {
-
         clusterer = ComplexBuilder<ItemKey>()
             .minClusteringZoom(9)
             .maxClusteringZoom(16)