+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# 3.21.0
+
+Release Date: 2025-03-20
+
+### 새로운 기능
+
+- 네이버 클라우드 플랫폼 키 ID 인증 방식 지원
+  - `NaverMapSdk.NcpKeyClient`
+  - `com.naver.maps.map.NCP_KEY_ID` 메타데이터
+- 커스텀 스타일링 기능 추가
+  - `NaverMapOptions.customStyleId()`
+  - `NaverMap#customStyleId`
+
 # 3.20.0
 
 Release Date: 2024-11-25